Add react-route-dom to switch between pages
Add react-route-dom to switch between 'home', 'about', and 'blog'.
Add additional CSS to Header.css to prevent the <Link> from altering the Header.js css.

The link was not added to the 'job' button because it is supposed to link to a specific section in the 'about' page. The about page needs to be built first before it can reference to that part of the 'about' page.